add common protocol binding interface
