+++ src/main/java/ch/unisg/ics/interactions/wot/td/bindings/BindingNotFoundException.java
@@ -1,0 +1,4 @@
+package ch.unisg.ics.interactions.wot.td.bindings;
+
+public class BindingNotFoundException extends RuntimeException {
+}
