+++ src/main/java/ch/unisg/ics/interactions/wot/td/bindings/Operation.java
@@ -1,0 +1,9 @@
+package ch.unisg.ics.interactions.wot.td.bindings;
+
+public interface Operation {
+
+  void setPayload();
+
+  Response execute();
+
+}
