+++ src/main/java/ch/unisg/ics/interactions/wot/td/bindings/Response.java
@@ -1,0 +1,4 @@
+package ch.unisg.ics.interactions.wot.td.bindings;
+
+public interface Response {
+}
