+++ src/main/java/ch/unisg/ics/interactions/wot/td/bindings/coap/TDCoapObserveRelation.java
@@ -1,4 +1,4 @@
-package ch.unisg.ics.interactions.wot.td.clients;
+package ch.unisg.ics.interactions.wot.td.bindings.coap;
 
 import org.eclipse.californium.core.CoapObserveRelation;
