+++ src/test/java/ch/unisg/ics/interactions/wot/td/bindings/coap/TDCoapRequestTest.java
@@ -1,9 +1,12 @@
-package ch.unisg.ics.interactions.wot.td.clients;
+package ch.unisg.ics.interactions.wot.td.bindings.coap;
 
 import ch.unisg.ics.interactions.wot.td.ThingDescription;
 import ch.unisg.ics.interactions.wot.td.affordances.ActionAffordance;
 import ch.unisg.ics.interactions.wot.td.affordances.Form;
 import ch.unisg.ics.interactions.wot.td.affordances.PropertyAffordance;
+import ch.unisg.ics.interactions.wot.td.bindings.coap.TDCoapHandler;
+import ch.unisg.ics.interactions.wot.td.bindings.coap.TDCoapRequest;
+import ch.unisg.ics.interactions.wot.td.bindings.coap.TDCoapResponse;
 import ch.unisg.ics.interactions.wot.td.io.TDGraphReader;
 import ch.unisg.ics.interactions.wot.td.schemas.*;
 import ch.unisg.ics.interactions.wot.td.vocabularies.COV;
