+++ src/test/java/ch/unisg/ics/interactions/wot/td/bindings/coap/TDCoapResponseTest.java
@@ -1,4 +1,4 @@
-package ch.unisg.ics.interactions.wot.td.clients;
+package ch.unisg.ics.interactions.wot.td.bindings.coap;
 
 import ch.unisg.ics.interactions.wot.td.schemas.*;
 import org.eclipse.californium.core.coap.Option;
