+++ src/test/java/ch/unisg/ics/interactions/wot/td/bindings/http/TDHttpResponseTest.java
@@ -1,4 +1,4 @@
-package ch.unisg.ics.interactions.wot.td.clients;
+package ch.unisg.ics.interactions.wot.td.bindings.http;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
